+++ packages/react-native/example/src/content.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useEffect} from 'react';
 import {View, Button, StyleSheet, Text, Platform} from 'react-native';
 import notifee, {
   AndroidImportance,
@@ -11,8 +11,16 @@ import {triggers} from './utils/triggers';
 type Props = {};
 
 export const Content: React.FC<Props> = () => {
+  useEffect(() => {
+    const unsubscribeNotifee = notifee.registerForegroundService(() => {
+      return new Promise(() => {});
+    });
+
+    return unsubscribeNotifee;
+  }, []);
+
   /* Change the notification payload */
-  const notification = notifications.basic;
+  const notification = notifications.image;
 
   /* Change the trigger type */
   const triggerType = triggers.timestamp;
@@ -23,8 +31,8 @@ export const Content: React.FC<Props> = () => {
     await notifee.createChannel({
       id: notification.android?.channelId || 'default',
       name: notification.android?.channelId || 'default',
-      importance: notification.android?.importance || AndroidImportance.DEFAULT,
-      visibility: notification.android?.visibility || AndroidVisibility.PRIVATE,
+      importance: notification.android?.importance || AndroidImportance.HIGH,
+      visibility: notification.android?.visibility || AndroidVisibility.PUBLIC,
       vibration: true,
       sound: notification.android?.sound || 'default',
     });
@@ -42,7 +50,7 @@ export const Content: React.FC<Props> = () => {
     await notifee.createChannel({
       id: notification.android?.channelId || 'default',
       name: notification.android?.channelId || 'default',
-      importance: notification.android?.importance || AndroidImportance.DEFAULT,
+      importance: notification.android?.importance || AndroidImportance.HIGH,
       visibility: notification.android?.visibility || AndroidVisibility.PRIVATE,
       vibration: true,
       sound: notification.android?.sound || 'default',