+++ packages/react-native/src/types/NotificationAndroid.ts
@@ -276,7 +276,7 @@ export interface NotificationAndroid {
    * The `fullScreenAction` property allows you to show a custom UI
    * in full screen when the notification is displayed.
    *
-   * View the [FullScreenAction](/react-native/docs/android/behaviour#full-screen-action) documentation to learn
+   * View the [FullScreenAction](/react-native/docs/android/behaviour#full-screen) documentation to learn
    * more.
    */
   fullScreenAction?: NotificationFullScreenAction;
@@ -1193,6 +1193,12 @@ export enum AndroidFlags {
    * This will be set for you by setting `loopSound`.
    */
   FLAG_INSISTENT = 4,
+
+  /**
+   * Prevents the notification from being canceled when the user clicks the Clear all button.
+   * This will be set for you by setting `ongoing`.
+   */
+  FLAG_NO_CLEAR = 32,
 }
 
 /**